split fm2profrunner tests
